refactor: extract raw-HTML scanner to scan::raw_html module (TB-8)
`find_raw_html`, `RawHtmlMatch`, and the tag-boundary helper move from
`validate.rs` into `crates/adoc-core/src/scan/raw_html.rs`. validate.rs
imports the scanner and shrinks to rule orchestration only — no more
130 lines of pure text-scanning logic mixed with the rule walk.

The new `scan/` submodule signals "pure text scanning, no AST awareness,
no diagnostic emission". Future text utilities (e.g. URL normalization for
the WHATWG follow-up issue) land alongside as `scan/url.rs` without
revisiting this layout.

Scanner unit tests move with the scanner; validator tests continue
exercising the rule via `validate_text` so the AST-walk integration
remains covered here.
