+++ crates/adoc-core/src/validate.rs
@@ -12,7 +12,8 @@
 use crate::ast::{BlockAst, PageAst, WorkspaceAst};
 use crate::diagnostic::{Diagnostic, DiagnosticCode, SourceSpan};
 use crate::inline::InlineSegment;
-use crate::source::{SourceFile, column_offset};
+use crate::scan::raw_html::find_raw_html;
+use crate::source::SourceFile;
 
 pub(crate) trait ValidationRule {
     fn check(&self, page: &PageAst, source: &SourceFile, sink: &mut Vec<Diagnostic>);
@@ -155,74 +156,6 @@ fn check_inlines(inlines: &[InlineSegment], sink: &mut Vec<Diagnostic>) {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct RawHtmlMatch {
-    start_column: u32,
-    end_column: u32,
-}
-
-fn find_raw_html(line: &str) -> Option<RawHtmlMatch> {
-    for (start_index, character) in line.char_indices() {
-        if character != '<' {
-            continue;
-        }
-
-        let is_tag_boundary = start_index == 0
-            || line[..start_index]
-                .chars()
-                .last()
-                .is_some_and(|character| character.is_whitespace());
-        if !is_tag_boundary {
-            continue;
-        }
-
-        let after_opening_bracket = &line[start_index + character.len_utf8()..];
-        let Some(tag_end) = raw_html_tag_end(after_opening_bracket) else {
-            continue;
-        };
-        let end_index = start_index + character.len_utf8() + tag_end;
-
-        return Some(RawHtmlMatch {
-            start_column: column_offset(&line[..start_index]),
-            end_column: column_offset(&line[..end_index]),
-        });
-    }
-
-    None
-}
-
-fn raw_html_tag_end(value: &str) -> Option<usize> {
-    let mut name_start = 0;
-    if value.starts_with('/') {
-        name_start = 1;
-    }
-
-    let first_character = value[name_start..].chars().next()?;
-    if !first_character.is_ascii_alphabetic() {
-        return None;
-    }
-
-    let mut name_end = name_start + first_character.len_utf8();
-    for character in value[name_end..].chars() {
-        if !character.is_ascii_alphanumeric() && character != '-' {
-            break;
-        }
-        name_end += character.len_utf8();
-    }
-
-    let next_character = value[name_end..].chars().next()?;
-    match next_character {
-        '>' => Some(name_end + 1),
-        '/' => value[name_end + 1..]
-            .starts_with('>')
-            .then_some(name_end + 2),
-        character if character.is_whitespace() => value[name_end..]
-            .find('>')
-            .map(|relative_index| name_end + relative_index + 1),
-        _ => None,
-    }
-}
-
 fn is_url_safe(url: &str) -> bool {
     if url.bytes().any(|byte| byte.is_ascii_whitespace()) {
         return false;
@@ -269,25 +202,10 @@ mod tests {
         diagnostics
     }
 
-    // --- predicate-level tests ---
-
-    #[test]
-    fn find_raw_html_matches_simple_block_tag() {
-        let m = find_raw_html("<div>x</div>").expect("expected match");
-        assert_eq!(m.start_column, 1);
-        assert_eq!(m.end_column, 6);
-    }
-
-    #[test]
-    fn find_raw_html_returns_none_for_inline_less_than() {
-        assert!(find_raw_html("Vec<String>").is_none());
-    }
-
-    #[test]
-    fn find_raw_html_skips_to_first_match_after_whitespace() {
-        let m = find_raw_html("hello <span>x</span>").expect("expected match");
-        assert_eq!(m.start_column, 7);
-    }
+    // Raw-HTML scanner unit tests live alongside the scanner in
+    // `crates/adoc-core/src/scan/raw_html.rs` (TB-8). Validator-level tests
+    // exercise the rule via `validate_text` so the integration with the AST
+    // walk and the per-block dispatch is covered here.
 
     #[test]
     fn is_url_safe_accepts_http_https_and_mailto() {
